Donator and funpk changes
I have added 2 maps, the funpk zone and donator area. both have commands.
::funpk and ::donatorarea
the donator command is temp, and the doors at the donator zone work and tele you into the areas but they can't tele you back into the safe zone at the donator area. The fun pk zone has teleports that do the same thing we will need to add the option to tele out for the teleports and doors. You may need the cache for the new maps to work.