Shops, Drops, & Molly Wops
In this update the Shops have been updated with Item Prices to balance out cost over the amount obtained through game play.

The hope tutorial has beed adjusted to the new home location.

Slayer masters are not able to be found in the home spawn inside of the castle as you walk in.

Some NPC Drops have been re-done.